restructure pipelines and add base class of pipeline
