+++ WeiboSpider/pipelines/UserInfoPipeline.py
@@ -3,23 +3,13 @@
 # @Time    : 2021/7/18 18:00
 # @Function:
 
-from WeiboSpider.database import DBConnector
 from WeiboSpider.items import UserInfoItem
+from WeiboSpider.pipelines.base import Pipeline
 
 
-class UserInfoPipeline:
+class UserInfoPipeline(Pipeline):
     def __init__(self, db_connector):
-        self.connector = db_connector
-
-    @classmethod
-    def from_crawler(cls, crawler):
-        return cls(db_connector=DBConnector())
-
-    def open_spider(self, spider):
-        self.db, self.client = self.connector.connect()
-
-    def close_spider(self, spider):
-        self.client.close()
+        super(UserInfoPipeline, self).__init__(db_connector)
 
     def process_item(self, item, spider):
         if isinstance(item, UserInfoItem):
