+++ WeiboSpider/pipelines/__init__.py
@@ -4,5 +4,6 @@
 # @Function:
 
 from WeiboSpider.pipelines.UserInfoPipeline import UserInfoPipeline
+from WeiboSpider.pipelines.TweetInfoPipeline import TweetInfoPipeline
 
-__all__ = ['UserInfoPipeline']
+__all__ = ['UserInfoPipeline', 'TweetInfoPipeline']
