+++ WeiboSpider/pipelines/base/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+# @Author  : CharesFuns
+# @Time    : 2021/7/18 18:59
+# @Function:
+
+from WeiboSpider.pipelines.base.Pipeline import Pipeline
+
+__all__ = ['Pipeline']
